chore(sdk): move code into src directory
