+++ packages/sdk/index.js
@@ -1,6 +1,6 @@
-import createMiddleware from './middleware';
-import * as actions from './actions';
-import SdkGet from './components/sdk-get';
+import createMiddleware from './src/middleware';
+import * as actions from './src/actions';
+import SdkGet from './src/components/sdk-get';
 
 const Sdk = {
   Get: SdkGet,
